Move datalayer.go to pkg/datalayer/types.go.
